Seperate unit tests for formatting_tools.c and marshalling_tools.c
